% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateSpines.R
\name{simulation3Dmesh}
\alias{simulation3Dmesh}
\title{3D virtual spine}
\usage{
simulation3Dmesh(newSpines, idx, iterations = 4)
}
\arguments{
\item{newSpines}{a matrix where each row is a virtual spine}

\item{idx}{is number of the row number of the spine to representate}
}
\value{
rgl object 3D mesh3d
}
\description{
Generate the 3D representation of the virtual spines
}
\examples{
newSpines <- spineSampling(model,nSpines=100,cluster=6,seed=1)
mesh <- simulation3Dmesh(newSpines,4,iterations=4)
shade3d(mesh,col="red")
shade3d(mesh,col=getColor(model$G,1))
}

