% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spineClustering.R
\name{spineClustering}
\alias{spineClustering}
\title{Clustering dendritic spines}
\usage{
spineClustering(csvSpines, numClusters = 15, scale = T)
}
\arguments{
\item{csvSpines}{string value which indicates the path to the csv file where are saved the morphological features of the spines}

\item{scale}{boolean value that indicates if the data should be standardized}

\item{G}{integer value that denotes the max number of clusters to check}
}
\value{
Mclust object that contains the mixture of multivariate Gaussians
}
\description{
Read a csv file with the features of the morphology of the dendritic spines. Cluster dendritic spines according to a mixture model of multivariate Gaussians.
}
\examples{
spineClustering(csvSpines = system.file("extdata", "spineDataset.csv", package = "spineSimulation"), numClusters = 15, scale = T)
}

