% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thresholdBelonging.R
\name{thresholdBelonging}
\alias{thresholdBelonging}
\title{Clustering dendritic spines}
\usage{
thresholdBelonging(model)
}
\arguments{
\item{a}{Mclust object}
}
\value{
num_spines_over_threshold a table with a probability of belonging lower than a threshold
}
\description{
Read a csv file with the features of the morphology of the dendritic spines. Cluster dendritic spines according to a mixture model of multivariate Gaussians.
}
\examples{
thresholdBelonging(model)
}

